Re-import scenarios into test folder
