+++ tests/appsec/iast/test_iast_sql_injection.py
@@ -2,7 +2,7 @@
 # This product includes software developed at Datadog (https://www.datadoghq.com/).
 # Copyright 2021 Datadog, Inc.
 
-from utils import BaseTestCase, interfaces, context, missing_feature, coverage, released, bug
+from utils import BaseTestCase, interfaces, context, missing_feature, coverage, released
 
 # Weblog are ok for nodejs/express4 and java/spring-boot
 @coverage.basic
