+++ tests/appsec/test_conf.py
@@ -3,7 +3,8 @@
 # Copyright 2021 Datadog, Inc.
 
 import pytest
-from utils import BaseTestCase, context, coverage, interfaces, released, missing_feature, irrelevant, rfc
+from utils import BaseTestCase, context, coverage, interfaces, released, missing_feature, irrelevant, rfc, scenario
+from tests.constants import PYTHON_RELEASE_GA_1_1
 from .waf.utils import rules
 
 
@@ -82,9 +83,67 @@ class Test_RuleSet_1_3_1(BaseTestCase):
 class Test_ConfigurationVariables(BaseTestCase):
     """ Configuration environment variables """
 
-    # some test are available in scenarios/
-
     def test_enabled(self):
         # test DD_APPSEC_ENABLED = true
         r = self.weblog_get("/waf/", headers={"User-Agent": "Arachni/v1"})
         interfaces.library.assert_waf_attack(r)
+
+    @irrelevant(library="ruby", weblog_variant="rack", reason="it's not possible to auto instrument with rack")
+    @missing_feature(
+        context.weblog_variant in ["sinatra14", "sinatra20", "sinatra21"],
+        reason="Conf is done in weblog instead of library",
+    )
+    @scenario("APPSEC_DISABLED")
+    def test_disabled(self):
+        """ test DD_APPSEC_ENABLED = false """
+        r = self.weblog_get("/waf/", headers={"User-Agent": "Arachni/v1"})
+        interfaces.library.assert_no_appsec_event(r)
+
+    @scenario("APPSEC_CUSTOM_RULES")
+    def test_appsec_rules(self):
+        """ test DD_APPSEC_RULES = custom rules file """
+        r = self.weblog_get("/waf", headers={"attack": "dedicated-value-for-testing-purpose"})
+        interfaces.library.assert_waf_attack(r, pattern="dedicated-value-for-testing-purpose")
+
+    @missing_feature(context.library < "java@0.113.0")
+    @missing_feature(context.library == "java" and context.weblog_variant == "spring-boot-openliberty")
+    @scenario("APPSEC_LOW_WAF_TIMEOUT")
+    def test_waf_timeout(self):
+        """ test DD_APPSEC_WAF_TIMEOUT = low value """
+        long_payload = "?" + "&".join(f"{k}={v}" for k, v in ((f"key_{i}", f"value{i}") for i in range(1000)))
+        r = self.weblog_get(f"/waf/{long_payload}", headers={"User-Agent": "Arachni/v1"})
+        interfaces.library.assert_no_appsec_event(r)
+
+    @missing_feature(context.library <= "ruby@1.0.0")
+    @missing_feature(context.library < f"python@{PYTHON_RELEASE_GA_1_1}")
+    @scenario("APPSEC_CUSTOM_OBFUSCATION")
+    def test_obfuscation_parameter_key(self):
+        """ test DD_APPSEC_OBFUSCATION_PARAMETER_KEY_REGEXP """
+
+        SECRET = "This-value-is-secret"
+
+        def validate_appsec_span_tags(span, appsec_data):  # pylint: disable=unused-argument
+            if SECRET in span["meta"]["_dd.appsec.json"]:
+                raise Exception("The security events contain the secret value that should be obfuscated")
+            return True
+
+        r = self.weblog_get("/waf", headers={"hide-key": f"acunetix-user-agreement {SECRET}"})
+        interfaces.library.assert_waf_attack(r, pattern="<Redacted>")
+        interfaces.library.add_appsec_validation(r, validate_appsec_span_tags)
+
+    @missing_feature(context.library <= "ruby@1.0.0")
+    @missing_feature(context.library < f"python@{PYTHON_RELEASE_GA_1_1}")
+    @scenario("APPSEC_CUSTOM_OBFUSCATION")
+    def test_obfuscation_parameter_value(self):
+        """ test DD_APPSEC_OBFUSCATION_PARAMETER_VALUE_REGEXP """
+
+        SECRET = "hide_value"
+
+        def validate_appsec_span_tags(span, appsec_data):  # pylint: disable=unused-argument
+            if SECRET in span["meta"]["_dd.appsec.json"]:
+                raise Exception("The security events contain the secret value that should be obfuscated")
+            return True
+
+        r = self.weblog_get("/waf", headers={"attack": f"acunetix-user-agreement {SECRET}"})
+        interfaces.library.assert_waf_attack(r, pattern="<Redacted>")
+        interfaces.library.add_appsec_validation(r, validate_appsec_span_tags)
