+++ tests/test_sampling_rates.py
